Tokenization for notesViewer
This commit is a rewrite of the entire logic that was used in the
notesViewer.
Now all the javascript that is required for the notesViewer is inserted
by the extension along with the tokens required for authentication. This
ensures that communication with the server is always done with an
authorized token and prevents stray traffic.

- Resolved #93